Rename clock_output -> wb_test
